Settle on adapter organization
