+++ lib/epoxi.ex
@@ -4,22 +4,16 @@ defmodule Epoxi do
   """
 
   alias Epoxi.Render
-  alias Epoxi.Parsing
-  alias Epoxi.Utils
+
+  defprotocol Adapter do
+    @doc "Adapters to implement to send a blocking email."
+    def send_blocking(context, email, message)
+  end
 
   def send_blocking(email, context) do
-    from_hostname = Parsing.get_hostname(email.from)
-
-    # TODO: pre-cache/lookup mx records for popular domains
-    {_priority, relay} =
-      Parsing.get_hostname(email.to)
-      |> Utils.mx_lookup()
-      |> List.first()
-
     message = Render.encode(email)
-    config = %{context.config | relay: relay, hostname: from_hostname}
 
-    case Epoxi.Adapter.send_blocking(config, email, message) do
+    case Adapter.send_blocking(context, email, message) do
       {:error, reason} ->
         IO.inspect(reason, label: :error_sending)
